Create WSGI application
For an application to talk to a web server, it must implement the WSGI
interface as specified by PEP 333. This interface is implemented as a
callable object such as a function, a method, a class, or an instance
with a __call__ method.

This object should accept two positional arguments and return the
response body as strings in an iterable.

The two arguments are:

- a dictionary with environment variables
- a callback function that will be used to send HTTP statuses and HTTP
  headers to the server